separate workflows into a separate file, and run from script.js
